LottoPrize 파일로 분리
