+++ src/test/java/lotto/domain/LottoResultTest.java
@@ -9,10 +9,10 @@ public class LottoResultTest {
     @Test
     public void  당첨금_검증() {
         LottoResult lottoResult = new LottoResult();
-        lottoResult.addResult(LottoMachine.getPrize(3));  // 5000
-        lottoResult.addResult(LottoMachine.getPrize(3));  // 5000
-        lottoResult.addResult(LottoMachine.getPrize(4));  // 50000
-        lottoResult.addResult(LottoMachine.getPrize(5));  // 1500000
+        lottoResult.addResult(LottoPrize.FIFTH);  // 5000
+        lottoResult.addResult(LottoPrize.FIFTH);  // 5000
+        lottoResult.addResult(LottoPrize.FOURTH);  // 50000
+        lottoResult.addResult(LottoPrize.THIRD);  // 1500000
 
         assertThat(lottoResult.getCountByMatchCount(3)).isEqualTo(2);
         assertThat(lottoResult.getCountByMatchCount(4)).isEqualTo(1);
